Add ExpandedMedia component and handle expanded media content
This commit adds the ExpandedMedia component to handle displaying expanded media content. The ExpandedMedia component accepts a post and a setExpandedMediaContent function as props. It renders an image or video based on the post data. If the post contains an image, it will be displayed using the Image component from the next/image package. The image is rendered within a fixed positioned div with a black background. The ExpandedMedia component also includes a button with a ChevronLeftIcon to close the expanded media view when clicked.

The UserPost component has been updated to include a div that wraps the PostImage and PostVideo components. Clicking on this div will call the setExpandedMediaContent function if the post has an image, setting the expanded media content to the current post.

The Home component now includes a state variable, expandedMediaContent, initialized as null, to keep track of whether an image should be expanded. If expandedMediaContent is truthy, the ExpandedMedia component is rendered with the expanded media content post. Otherwise, the UserPost components are rendered normally.

A new type, Post, has been created in the types/Post.ts file to represent a post object returned from the API. The Post type is based on the RouterOutputs type from the utils/api file.

The PostInfo type has been deleted as it is no longer used in the codebase.